app: apply all changes on all events
Towards https://github.com/giantswarm/giantswarm/issues/5282
Towards https://github.com/giantswarm/giantswarm/issues/5461

This resource is generic. We actually want to be able to delete app CRs
on update events.
